refactor form logic into custom hook
