+++ pages/add.js
@@ -1,45 +1,25 @@
 import styled from 'styled-components'
-import { useRouter } from 'next/router'
-import { nanoid } from 'nanoid'
-import { useState } from 'react'
+import useSpendForm from '../hooks/useSpendForm'
 
 export default function Add({ addSpend }) {
-  const router = useRouter()
-
-  const [moneyFor, setMoneyFor] = useState('')
-  const [amountOfMoney, setAmountOfMoney] = useState(0)
-  const [payDate, seetPayDate] = useState('')
-
-  const [splits, setSplit] = useState([])
-
-  function addNewSplit() {
-    const newSplit = {
-      id: nanoid(),
-      nameValue: 'martin',
-      number: 0,
-      moneyValue: 'martin',
-    }
-    setSplit([...splits, newSplit])
-  }
-
-  function deleteSplit(id) {
-    const updatedSplit = splits.filter((split) => split.id !== id)
-    setSplit(updatedSplit)
-  }
-
-  function handleSplitChangeByType(id, value, type) {
-    const updatedSplit = splits.map((split) => {
-      if (split.id === id) {
-        return { ...split, [type]: value }
-      } else {
-        return split
-      }
-    })
-    setSplit(updatedSplit)
-  }
+  const {
+    moneyFor,
+    setMoneyFor,
+    amountOfMoney,
+    setAmountOfMoney,
+    payDate,
+    setPayDate,
+    splits,
+    addNewSplit,
+    deleteSplit,
+    handleSplitChangeByType,
+    resetAll,
+    saveSpend,
+    dontSaveAndGoBack,
+    checkIsReadyToSave,
+  } = useSpendForm(addSpend)
 
   const isReadyToSave = checkIsReadyToSave()
-
   return (
     <Form onSubmit={saveSpend}>
       <label>Ausgabe für:</label>
@@ -61,7 +41,7 @@ export default function Add({ addSpend }) {
       <input
         type="date"
         value={payDate}
-        onChange={({ target: { value } }) => seetPayDate(value)}
+        onChange={({ target: { value } }) => setPayDate(value)}
       ></input>
 
       <h2>Aufteilung</h2>
@@ -124,42 +104,6 @@ export default function Add({ addSpend }) {
       </button>
     </Form>
   )
-
-  function resetAll() {
-    setMoneyFor('')
-    setAmountOfMoney(0)
-    seetPayDate('')
-    setSplit([])
-  }
-
-  function saveSpend(event) {
-    event.preventDefault()
-
-    const newSpend = {
-      id: nanoid(),
-      spendFrom: 'martin',
-      moneyFor,
-      amountOfMoney,
-      payDate,
-      splits,
-    }
-
-    addSpend(newSpend)
-    router.push('/')
-  }
-
-  function dontSaveAndGoBack() {
-    router.push('/')
-  }
-
-  function checkIsReadyToSave() {
-    return (
-      moneyFor.trim().length > 0 &&
-      amountOfMoney > 0 &&
-      payDate.length > 0 &&
-      splits.length > 0
-    )
-  }
 }
 
 const SplitItem = styled.li``
